Instance Generators

QuadraticAssignmentProblem.uniform_matrix — Function

Generate an N×N matrix with uniformly random entries in [0,1].

QuadraticAssignmentProblem.metric_matrix — Function

Generate a random N×N matrix with entries satisfying the triangle inequality.

QuadraticAssignmentProblem.zeroone_matrix — Function

Generate a random 0-1 N×N matrix with edge probability p.

QuadraticAssignmentProblem.generate_qap — Function

Generate a pair of N×N matricies with known optimal permutation using Algorithm 4 from "Generating Quadratic Assignment Test Problems with Known Optimal Permutations" by Li and Pardalos.