Solvers

QuadraticAssignmentProblem.qap_exact — Function

Solve the QAP using quadratic integer progamming.

QuadraticAssignmentProblem.qap_linearization — Function

Approximately solve the QAP by linearizing the problem and solving the resulting linear integer program.

QuadraticAssignmentProblem.qap_lprounding — Function

Approximately solve the QAP by linearizing and relaxing the problem and solving the resulting LIP with linear programming and rounding via the Hungarian algorithm.

QuadraticAssignmentProblem.qap_random — Function

Approximately solve the QAP through random sampling.

QuadraticAssignmentProblem.qap_mms — Function

Run the algorithm from "Maximum Quadratic Assignment Problem: Reduction from Maximum Label Cover and LP-based Approximation Algorithm" by Makarychev, Manokaran, and Sviridenko (MMS) to approximately solve the QAP.

QuadraticAssignmentProblem.qap_ns — Function

Approximate the maxQAP using the complete algorithm from "On the Maximum Quadratic Assignment Problem" by Nagarajan and Sviridenko (NS).

Approximate the 0-1 QAP using the complete algorithm from "On the Maximum Quadratic Assignment Problem" by Nagarajan and Sviridenko (NS).

QuadraticAssignmentProblem.faq — Function

Perform the FAQ algorithm from "Fast Approximate Quadratic Programming for Graph Matching" to approximately solve the QAP.